+++ packages/polaris-viz/src/components/ChartSkeleton/index.ts
@@ -1,3 +1,9 @@
 export {ChartSkeleton} from './ChartSkeleton';
-export type {ChartSkeletonProps} from './ChartSkeleton';
+export type {
+  DonutSkeletonProps,
+  FunnelSkeletonProps,
+  SimpleBarSkeletonProps,
+  SparkSkeletonProps,
+  SimpleNormalizedSkeletonProps,
+} from './ChartSkeleton';
 export type {SkeletonType} from './ChartSkeleton';